% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryColumnFrequency.R
\name{buildQueryColumnFrequency}
\alias{buildQueryColumnFrequency}
\title{buildQueryColumnFrequency}
\usage{
buildQueryColumnFrequency(conn.info, ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{...}{Other parameters}
}
\value{
query column, count(*) from table
}
\description{
buildQueryColumnFrequency
}
