% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryColumnFrequency.R
\name{buildQueryColumnFrequency.sqlserver}
\alias{buildQueryColumnFrequency.sqlserver}
\title{buildQueryColumnFrequency.sqlserver}
\usage{
\method{buildQueryColumnFrequency}{sqlserver}(conn.info, schema, table, column,
  limit.freq.values, query.filter, ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{schema}{Table Schema}

\item{table}{Table Name}

\item{column}{Column profiled}

\item{limit.freq.values}{Number of frequency values detailed}

\item{query.filter}{Filter applied to the profile}

\item{...}{Other parameters}
}
\value{
query column, count(*) from table
}
\description{
buildQueryColumnFrequency.sqlserver
}
