% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryColumnMetadata.R
\name{buildQueryColumnMetadata.sqlite}
\alias{buildQueryColumnMetadata.sqlite}
\title{buildQueryColumnMetadata.sqlite}
\usage{
\method{buildQueryColumnMetadata}{sqlite}(conn.info, table)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{table}{Table Name}
}
\value{
query columns' metadata
}
\description{
buildQueryColumnMetadata.sqlite
}
