% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryColumnMetadata.R
\name{buildQueryColumnMetadata.sqlserver}
\alias{buildQueryColumnMetadata.sqlserver}
\title{buildQueryColumnMetadata.sqlserver}
\usage{
\method{buildQueryColumnMetadata}{sqlserver}(conn.info, schema, table, database,
  ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{schema}{Table Schema}

\item{table}{Table Name}

\item{database}{Database Name}

\item{...}{Other params}
}
\value{
query columns' metadata
}
\description{
buildQueryColumnMetadata.sqlserver
}
