% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryColumnMetadata.R
\name{buildQueryColumnMetadata.teradata}
\alias{buildQueryColumnMetadata.teradata}
\title{buildQueryColumnMetadata.teradata}
\usage{
\method{buildQueryColumnMetadata}{teradata}(conn.info, schema, table)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{schema}{Table Schema}

\item{table}{Table Name}
}
\value{
query columns' metadata
}
\description{
buildQueryColumnMetadata.teradata
}
