% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryColumnStats.R
\name{buildQueryColumnStats.sqlite}
\alias{buildQueryColumnStats.sqlite}
\title{buildQueryColumnStats.sqlite}
\usage{
\method{buildQueryColumnStats}{sqlite}(conn.info, schema, table, column,
  query.filter, ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{schema}{Table Schema}

\item{table}{Table Name}

\item{column}{Column profiled}

\item{query.filter}{Filter applied to the profile}

\item{...}{Other parameters}
}
\value{
query count(distinct column) from table
}
\description{
buildQueryColumnStats.sqlite
}
