% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryCountNull.R
\name{buildQueryCountNull.sqlserver}
\alias{buildQueryCountNull.sqlserver}
\title{buildQueryCountNull.sqlserver}
\usage{
\method{buildQueryCountNull}{sqlserver}(conn.info, schema, table, column,
  query.filter, ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{schema}{Table Schema}

\item{table}{Table Name}

\item{column}{Column profiled}

\item{query.filter}{Filter applied to the profile}

\item{...}{Other parameters}
}
\value{
query select count(*) where collumn is null
}
\description{
buildQueryCountNull.sqlserver
}
