% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryCountTotal.R
\name{buildQueryCountTotal.sqlite}
\alias{buildQueryCountTotal.sqlite}
\title{buildQueryCountTotal.sqlite}
\usage{
\method{buildQueryCountTotal}{sqlite}(conn.info, schema, table, query.filter,
  ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{schema}{Table Schema}

\item{table}{Table Name}

\item{query.filter}{Filter applied to the profile}

\item{...}{Other parameters}
}
\value{
query count(*) from table
}
\description{
buildQueryCountTotal.sqlite
}
