% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryProfileColumnFormatFrequency.R
\name{buildQueryProfileColumnFormatFrequency}
\alias{buildQueryProfileColumnFormatFrequency}
\title{buildQueryProfileColumnFormatFrequency}
\usage{
buildQueryProfileColumnFormatFrequency(conn.info, ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{...}{Other parameters}
}
\value{
queries column format frequency from table
}
\description{
buildQueryProfileColumnFormatFrequency
}
