% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryProfileColumnFormatFrequency.R
\name{buildQueryProfileColumnFormatFrequency.sqlite}
\alias{buildQueryProfileColumnFormatFrequency.sqlite}
\title{buildQueryProfileColumnFormatFrequency.sqlite}
\usage{
\method{buildQueryProfileColumnFormatFrequency}{sqlite}(conn.info, schema,
  table, column, column.datatype, query.filter, ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{schema}{Table Schema}

\item{table}{Table Name}

\item{column}{Column profiled}

\item{column.datatype}{Column datatype defined at the rdbs}

\item{query.filter}{Filter applied to the profile}

\item{...}{Other parameters}
}
\value{
queries column format frequency from table
}
\description{
buildQueryProfileColumnFormatFrequency.sqlite
}
