% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryProfileColumnFormatFrequency.R
\name{buildQueryProfileColumnFormatFrequency.sqlserver}
\alias{buildQueryProfileColumnFormatFrequency.sqlserver}
\title{buildQueryProfileColumnFormatFrequency.sqlserver}
\usage{
\method{buildQueryProfileColumnFormatFrequency}{sqlserver}(conn.info, schema,
  table, column, query.filter, ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{schema}{Table Schema}

\item{table}{Table Name}

\item{column}{Column profiled}

\item{query.filter}{Filter applied to the profile}

\item{...}{Other parameters}
}
\value{
queries column format frequency from table
}
\description{
buildQueryProfileColumnFormatFrequency.sqlserver
}
