% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeConnection.R
\name{closeConnection}
\alias{closeConnection}
\title{closeConnection}
\usage{
closeConnection(conn)
}
\arguments{
\item{conn}{Connection created at \code{\link{connectDB}}}
}
\value{
\code{TRUE} if succeeded at closing connection
}
\description{
Disconnects from database using odbc::dbDisconnect
}
