% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectDB.R
\name{connectDB.default}
\alias{connectDB.default}
\title{connectDB.default}
\usage{
\method{connectDB}{default}(conn.info, ...)
}
\arguments{
\item{conn.info}{Connection info created at \code{\link{prepareConnection}}}

\item{...}{Other parameters}
}
\value{
\code{connection} to database
}
\description{
Connects to database using \code{\link[odbc]{dbConnect}}
}
