% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableColumns.R
\name{getTableColumns}
\alias{getTableColumns}
\title{getTableColumns}
\usage{
getTableColumns(conn.info, schema, table)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{schema}{Table schema}

\item{table}{Table name}
}
\value{
data frame containing the columns' metadata
}
\description{
Issues query against the RDBS to retrieve information about each column
of the table. Name, type, length, precision, etc.
}
