% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.profile}
\alias{print.profile}
\title{Print method}
\usage{
\method{print}{profile}(x, ...)
}
\arguments{
\item{x}{profile object}

\item{...}{other parameters}
}
\value{
printed profile
}
\description{
Print method
}
