% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileColumnFormat.R
\name{profileColumnFormat}
\alias{profileColumnFormat}
\title{profileColumnFormat}
\usage{
profileColumnFormat(conn.info, column, column.datatype, schema, table,
  count.total, query.filter, format.show.percentage)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{column}{Column name that will be profiled}

\item{column.datatype}{Column datatipe}

\item{schema}{Table schema}

\item{table}{Table name}

\item{count.total}{Number of rows to be profiled}

\item{query.filter}{Filter applied to the table, when profilling}

\item{format.show.percentage}{Threshold considered when showing formats'
percentages}
}
\value{
Data Frame containing columns' metadata
}
\description{
Profiles column based on its format, using masking strategy.
X = char, 9 = digit, S = symbol
}
