% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runProfile.R
\name{runProfile}
\alias{runProfile}
\title{Profile all columns from ODBC table or dataframe}
\usage{
runProfile(conn.info, schema = NULL, table, is.parallel = TRUE,
  count.nodes, query.filter = NA, format.show.percentage = 0.03)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{schema}{Table schema}

\item{table}{Table name}

\item{is.parallel}{Boolean that indicates if profile will run in parallel. Default TRUE.}

\item{count.nodes}{Number of nodes used when is.parallel = TRUE}

\item{query.filter}{Filter applied to the table, when profilling}

\item{format.show.percentage}{Threshold considered when showing formats percentages}
}
\value{
profile results for the table/dataframe
}
\description{
Profiles tables and dataframes (collecting statistics and informative summaries
about that data): max, min, avg, sd, nulls, distinct values, data patterns, data/format frequencies.
}
