% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.profile}
\alias{summary.profile}
\title{Override summary function}
\usage{
\method{summary}{profile}(object, ...)
}
\arguments{
\item{object}{Profile object}

\item{...}{other parameters}
}
\value{
data.frame with summary information
}
\description{
Override summary function
}
