% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryColumnFrequency.R
\name{buildQueryColumnFrequency}
\alias{buildQueryColumnFrequency}
\title{buildQueryColumnFrequency}
\usage{
buildQueryColumnFrequency(conn.info, ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection()}}}

\item{...}{other parameters}
}
\value{
query count frequency of values
}
\description{
buildQueryColumnFrequency
}
