% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryColumnMetadata.R
\name{buildQueryColumnMetadata}
\alias{buildQueryColumnMetadata}
\title{buildQueryColumnMetadata}
\usage{
buildQueryColumnMetadata(conn.info, ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection()}}}

\item{...}{specific parameters for each class}
}
\value{
query to retrieve columns' metadata
}
\description{
buildQueryColumnMetadata
}
