% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryColumnStats.R
\name{buildQueryColumnStats}
\alias{buildQueryColumnStats}
\title{buildQueryColumnStats}
\usage{
buildQueryColumnStats(conn.info, ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection()}}}

\item{...}{other parameters}
}
\value{
query min and max values
}
\description{
buildQueryColumnStats
}
