% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryCountNull.R
\name{buildQueryCountNull}
\alias{buildQueryCountNull}
\title{buildQueryCountNull}
\usage{
buildQueryCountNull(conn.info, ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{...}{other parameters}
}
\value{
query count(null)
}
\description{
buildQueryCountNull
}
