% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryProfileColumnFormatFrequency.R
\name{buildQueryProfileColumnFormatFrequency}
\alias{buildQueryProfileColumnFormatFrequency}
\title{buildQueryProfileColumnFormatFrequency}
\usage{
buildQueryProfileColumnFormatFrequency(conn.info, ...)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{...}{other parameters}
}
\value{
query format frequency (X = letter, 9 = digit, b = white space)
}
\description{
buildQueryProfileColumnFormatFrequency
}
