% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeConnection.R
\name{closeConnection}
\alias{closeConnection}
\title{Disconnnects from database using \code{\link{odbc::dbDisconnect}}
#'}
\usage{
closeConnection(conn)
}
\arguments{
\item{conn}{Connection created at \code{\link{connectDB}}}
}
\value{
\code{TRUE} if succeeded at closing connection
}
\description{
Disconnnects from database using \code{\link{odbc::dbDisconnect}}
#'
}
