% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileColumn.R
\name{profileColumn}
\alias{profileColumn}
\title{Title}
\usage{
profileColumn(conn.info, schema, table, column, column.datatype, query.filter,
  limit.freq.values = 30)
}
\arguments{
\item{limit.freq.values}{}
}
\description{
Title
}
