% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runProfile.R
\name{runProfile}
\alias{runProfile}
\title{Profiles datasets (collecting statistics and informative summaries
about that data) on data frames and ODBC tables: max, min, avg, sd, nulls,
distinct values, data patterns, data/format frequencies.}
\usage{
runProfile(conn.info, schema = NULL, table, is.parallel = TRUE,
  count.nodes = 5, query.filter = NA)
}
\arguments{
\item{conn.info}{Connection info created with \code{\link{prepareConnection}}}

\item{schema}{Table schema}

\item{table}{Table name}

\item{is.parallel}{Boolean that indicates if profile will run in parallel}

\item{count.nodes}{Number of nodes used when is.parallel = TRUE}

\item{query.filter}{Filter applied to the table, when profilling}
}
\value{
profile results for the table
}
\description{
Profiles datasets (collecting statistics and informative summaries
about that data) on data frames and ODBC tables: max, min, avg, sd, nulls,
distinct values, data patterns, data/format frequencies.
}
\examples{
conn.info <- prepareConnection(db.vendor = "teradata",
   dsn = "ODBC_MYDB", user = "myuser", passwd = "mypasswd")
p <- runProfile(conn.info, "table_schema", "table_name")
}
