% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.profile}
\alias{summary.profile}
\title{Shows a summary of the data profile}
\usage{
\method{summary}{profile}(object, ...)
}
\arguments{
\item{object}{Profile object}

\item{...}{}
}
\value{
data.frame with summary information
}
\description{
Shows a summary of the data profile
}
